+++ CHANGELOG.md
@@ -2,9 +2,20 @@
 
 *Please check the [releases](https://github.com/cars10/elasticvue/releases) page if you need version release dates.*
 
+## 0.22.0 (unreleased)
+
+* [feat]: change searching to use elasticsearch pagination+sort. **this limits the filter to the current page.** [#20][20]
+* [feat]: can click on index aliases on index overview to directly search
+* [fix]: fix showing index info/stats not always showing the correct result in modal
+* remove the deprecated [elasticsearch](https://www.npmjs.com/package/elasticsearch) library and api browser page. see [#22][22] for details
+
+[20]: https://github.com/cars10/elasticvue/issues/20
+[22]: https://github.com/cars10/elasticvue/issues/22
+
+
 ## 0.21.0
 
-* adds support for hosts with path/subdirectory where elasticsearch is not available at the url root, for example `example.com/elasticsearch` ([#19][i19])
+* [fix]: adds support for hosts with path/subdirectory where elasticsearch is not available at the url root, for example `example.com/elasticsearch` ([#19][i19])
 
 [i19]: https://github.com/cars10/elasticvue/issues/19